requirements.txt
numpy==1.20.3
tqdm==4.62.3
fire==0.4.0
torch==1.10.0
transformers[speech, sentencepiece]==4.12.3
pyyaml~=6.0
easydict~=1.9
torchaudio~=0.10.0
librosa==0.8.1
gradio~=2.4.5
onnxruntime~=1.9.0
soundfile~=0.10.3
onnx~=1.10.2
scipy~=1.7.2
